Use utils.jinja instead of class in utils
